Explicitly encode state machine for ops transitions
Summary:
This is where it finally gets interesting: `advanceReadOperation_` and `advanceWriteOperation_` replace the very complicated `triggerReadyCallbacks_` and `checkForMessagesDoneCollectingTensorDescriptors_` and a whole bunch of other checks spread throughout the code. It's not like those two new methods are straightforward, but at least all the "business logic" is concentrated in a single place and expressed as explicitly as possible.

The structure of the pipe now has become like this:
- there are the entry points, `readDescriptor` and `write`, where a new operation is created
- there are the `advanceFooOperation` method that encode the state machine
- there are the "action" methods, which perform a transition on the operations: these could be about I/O (`readPayloads`, `sendTensors`, ...) or about signaling something back to the user (`callFooCallback`)
- there are the "reaction" methods, which are called from inside callbacks. Note that, in a sense, `read` is a reaction, because it's the user "calling back in" to the pipe to inform us that the allocation has been completed.

Once this diff lands I want to move some code around in this file to reorder methods according to the above classification and add comments to that effect.

Reviewed By: beauby

Differential Revision: D21866467

fbshipit-source-id: 44f3bff6f3b8c8e75c257814163aae68a4a62827