Merge three classes for incoming messages into one
Summary:
The resulting class is just the union of the fields of the previous classes (deduplicated and disambiguated).

Having three separate classes, whose objects were stored in three separate deques, also meant that we were to, say, access the first unallocated message in constant time, as it was the front of one of these deques. Now, however, that message would just be somewhere in the middle of the single deque. To address this, we store the sequence numbers that corresponds to where these "cutpoints" where before, so that using these sequence numbers we can still access in constant time. So, for example, popping from one deque and pushing on another just becomes increasing the sequence number of that cutpoint.

Reviewed By: beauby

Differential Revision: D21866468

fbshipit-source-id: 01e108b2ba28f1db320c5c4a9a761888f68f395d