Expose name of remote context of pipe to users
Summary: In the PyTorch agent, when we open a outgoing pipe, we know the name of the worker that we think we're connecting to, and thus can use that name in all the logs to make them more informative. However, we don't know the name of the remote worker for *incoming* pipes. In fact, TensorPipe knows that name, as it is transferred as part of the "hello" packet, and TensorPipe uses that name to give the incoming pipe a meaningful name (i.e., `bar[l0]_from_foo`). What this diff does is allow the user (i.e., the agent) to retrieve that name from the pipe and thus be able to use it in their own logging.

Reviewed By: jiayisuse

Differential Revision: D21932985

fbshipit-source-id: 69996370de8962558630a524575ef94de8135723