Use context name rather than ID in logs
Summary: In log messages we were using names like `12345:c0.p1` or `67890:c0.l1.p2` to identify pipes (which combined the PID and some sequence numbers). While these names are unique, and hierarchical, they are not at all informative. So in this diff I am replacing them with some more semantically meaningful names: `foo_to_bar` and `bar[l0]_from_foo`, where `foo` and `bar` are the names that are provided by the user to the constructor and the `connect` method of the pipe (in PyTorch these correspond to the worker names, so in the tests this would be `worker0`, for example). These names still degrade to something similar to the old names if the user doesn't provide any names. This makes debugging much easier and, in fact, we lose little information because thanks to the previous diff (D21932986) we're already printing the PID separately.

Reviewed By: jiayisuse

Differential Revision: D21932990

fbshipit-source-id: 6fdc5daeb14badf2c6c694170c87f84c2e96eae2