Fix sending of remote worker name in pipe
Summary: I guess I made a last-minute change in D21932985 which I didn't double-test, where I moved the name into the ID instead of copying it. We keep ID and name separate because we want to use the name (if given) for both local logging and to send to the remote end, but we don't want to send the autogenerated ID to the remote end. The result, however, ended up being that the name was always empty and thus never sent to the remote side.

Reviewed By: jiayisuse

Differential Revision: D21953204

fbshipit-source-id: 3652a80d8909c29cd7d77f3647ba8ecba1fb5ece