Fix race condition between firing epoll handlers and changing their registration
Summary:
We had a race condition where we would attempt to fire a handler for an epoll event after that handler had already been unregistered, just because the events were received on one thread (the loop thread) but then had to be processed on another one (the reactor one) and anything could happen in the meanwhile. This was typically not a problem, because if the handler was gone we'd just do nothing. However, it became occasionally a problem when another socket with the same file descriptor value was registered: in that case we'd fire the new handler with the old events (which typically included some error).

The solution is to introduce an additional level of indirection, where instead of storing the fd on the epoll event, we store a more unique identifier (a "record"), which allows us to tell apart two versions of the same file descriptor, or even two handlers for the same file descriptor.

Reviewed By: beauby

Differential Revision: D21996947

fbshipit-source-id: c0ec3b07da68eb9cfc134081c7d0b3258d948564