Hold shared_ptrs instead of weak_ptrs to epoll handlers
Summary: I think the weak_ptrs were a legacy of the era when we didn't have a good threading model for the ownership of objects. Now we do, and both the connection and the listener (the two objects that actually register a handle with epoll) do unregister it when handling an error, and this error handling also covers closing the listener/connection, which is what happens when the "main" shared_ptr (the one held by the user) is destroyed. So it should be safe to use shared_ptrs. In fact, I consider it to be safer: while, contrary to weak_ptrs, it may cause memory leaks, those leaks would be signs that our logic is off, and thus would point to a broader issue, which we'd rather be informed about instead of silencing it.

Reviewed By: beauby

Differential Revision: D21996946

fbshipit-source-id: e1e0e014e119d2dbb4a3ef115874af80592e3ca7