Remove FunctionEventHandler from SHM loop
Summary: In D20941788 we had already removed the so-called "Monitor" from SHM's loop, and only put in the tests. We're now doing the same with the FunctionEventHandler, which was only used once, in order to register the eventfd inside the epoll loop itself. In fact, we can special-case the eventfd, and therefore do without the FunctionEventHandler. Apart from removing all that machinery for that one simple usage, it also allows us to add an assert on the fact that the registerDescriptor and unregisterDescriptor methods are only called from within the reactor loop, which is an important property that we'll need in later diffs.

Reviewed By: beauby

Differential Revision: D21996950

fbshipit-source-id: ae60760b4ca8cac2c5c17cbfea7449a9391e7c96