Have epoll defer to loop rather than using reactor token
Summary: In D21762966 we have introduced a separate codepath for deferring functions to the reactor that doesn't write to the reactor's ringbuffer. What this means is that the ringbuffer is only needed for *other processes* to cause a reaction in this process. The epoll loop however was still using a ringbuffer token to defer to the reactor. This diff fixes that.

Reviewed By: beauby

Differential Revision: D21996948

fbshipit-source-id: f1791aa882db82ca0369a2fdbd2c2a8cf0c6e40a