Decouple reactor from loop
Summary: For historical reasons, the reactor was a subfield of the loop. After the recent changes, it became clear that in fact it was the loop that was a customer of the reactor. So here we "promote" the reactor by taking it out from the loop and making it a field of the context impl, i.e., a "sibling" of the loop.

Reviewed By: beauby

Differential Revision: D22015790

fbshipit-source-id: a996168551d68503286887a12b417b299da7757b