Fix MPT in case of tensors with fewer bytes than num of lanes
Summary:
When registering MPT in PyTorch, using 16 lanes, found that it was segfaulting. It turned out that the cause was that I was calling `torch.tensor((N,))` instead of `torch.empty((N,))` and thus getting a much smaller tensor than I thought. This caused some of the chunks of the tensor to be empty. The MPT channel calls the user callback and deletes the operation as soon as all data has been transferred, which means it would do so even when some chunks were still pending, if those chunks had a length of zero. Thus, when those chunks finished, the reference to the operation that they were holding was stale, and this caused a crash.

There is no difference, in the general case, between waiting for all data to have been transferred and waiting for all chunks to have been transferred. So in this diff I am changing the code to just keep a count of the outstanding chunks.

I wrote two tests that would have caught this issue earlier. One of them also found an issue in the XTH channel, which I am fixing here too.

Reviewed By: beauby

Differential Revision: D22159537

fbshipit-source-id: 5a5ddd1317a82dcc714007cea8d7eb1ca8a47216