Define preprocessor flags in headers rather than through CMake (#191)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/191

Two of our backends, the SHM transport and the CMA channel, are Linux-specific and are thus built optionally depending on the platform. In order for downstream users to detect their availability we defined two preprocessor flags through CMake. This was brittle and we had in fact messed it up (we had two CMake files, a main one and a PyTorch one, and we forgot these flags in the latter). It would be more robust to define them in a header file, so that it's independent on the build system (well, the generation of the file will be dependent on the build system, but once the file has been generated it's not anymore).

Differential Revision: D22797910

fbshipit-source-id: c6b019007c8648ba60a78c7a115f144818f20dde