Export CMake targets from top-level directory (#194)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/194

At some point I think I have seen that the location in which that install command is triggered affects the location of the installed file. For us, doing it from the top-level doesn't change anything, so why not play it safe? Also, this is the same that Gloo does (and Gloo also tries to keep most of its CMake logic inside its gloo/ subdir).

Test Plan: Imported from OSS

Reviewed By: malfet

Differential Revision: D22928858

Pulled By: lw

fbshipit-source-id: c267ac7556c01096671910d305e12d51d1e56a61