Export submodule libuv inside TensorPipe's targets (#195)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/195

That target is a hack we need in order to be able to export the libuv target. I think it would be weird for users to install TensorPipe and then find a weird thing called libuv-unofficial installed on their machine. Why not keep it under our umbrella to make attribution easier? A later PR in this stack will also change the name of the target itself (currently `_uv_a`) as well.

Test Plan: Imported from OSS

Reviewed By: malfet

Differential Revision: D22928856

Pulled By: lw

fbshipit-source-id: c8406499e76e8659528b6b7433d6e316713c650a