Allow downstream projects to customize install location (#197)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/197

It makes perfect sense for us to install in the standard location for libraries in the system. However, this sometimes turns out to be `lib64`, while PyTorch hardcodes such a location as `lib`. Not sure it matters, and if anything it should be PyTorch to fix this, but let's be nice and allow users to specify a different location if they want.

Test Plan: Imported from OSS

Reviewed By: malfet

Differential Revision: D22928855

Pulled By: lw

fbshipit-source-id: ee3c12ed251a86e71c81ce5f0108b32318e0649a