Replace protobuf with libnop in connections
Summary:
This is the big change, where we swap out protobuf for libnop in the biggest usecases, namely the pipe and the channels. Luckily all these changes are identical so although this consists in many lines the complexity of this change is limited.

What we do is:
- change the interface of connections to provide a specialization of read/write for NopHolders, rather than for protobuf Messages.
- convert all the proto definitions to C++ structs, defined in new nop_types.h files.
- go through all usages of the protobufs and all callsites of read/write and update them.

This change leaves behind some proto defs which are still being used by some unit tests, as these tests will be ported in later commits and those protos will be removed then.

It also doesn't reimplement yet the specialized codepath that the SHM has for protobuf messages: it is left there, albeit "cut off", which means that SHM will fall back to the generic implementation and be slower. This will be temporary as in the next commit we'll port that specialized code and re-attach it.

Reviewed By: heiner

Differential Revision: D22763737

fbshipit-source-id: d44be13e827822662a0b0395aff05ece478dbc51