Promote SHM's fd & socket RAII wrappers to project-wide helpers
Summary:
We had RAII wrappers for file descriptors and sockets in SHM which were quite helpful. In some places, where we couldn't make use of the Fd wrapper, we were "releasing" file descriptors and then having an other object re-own them, and in between these steps we were dealing with raw integers. Just like when dealing with raw pointers, that makes me queasy. As for the Socket wrapper, it had a very helpful method for sending fds over a unix domain socket, which we wanted to use in the tests, and in order to do so we had some tests depend on the SHM transport, which was weird.

I think it's nicer to make these helpers available to the entire codebase, so that we can use the Fd wrapper throughout (note how we can now delete its `release` method! and note how the Segment now doesn't close its file descriptor on its own but simply holds a Fd which does it on its own) and we can use the Socket methods easily.

Reusability of these components will become more interesting shortly, with the InfiniBand transport.

Reviewed By: beauby

Differential Revision: D23567124

fbshipit-source-id: ca33325a0da4aeb03cde24bb1912049911aa16f2