Clean up ringbuffer consumer
Summary:
With the recent migration from protobuf to libnop we've simplified the set of methods we require the ringbuffer consumer/producer to provide. We also came up with a nice low-level general purpose interface to access the ringbuffer: accessing a contiguous region through either one ptr+len or two (if it wraps).

This diff takes the chance we now have to simplify the ringbuffer's methods, starting by the consumer. It makes the "access a contiguous region" method the foundation of everything else: that's now the only place where the ringbuffer logic resides, and everything else builds on top of it like a stack. Note that "everything else" now boils down to two methods: one that, in addition, copies the data out of the ringbuffer into a user-provided buffer, and another one that, on top, deals with transactions for you.

The advantage of this change is that we now have a very consistent experience of the ringbuffer, no matter what one tries to do with it (e.g., we still had some methods that didn't support zero-sized lengths). It also helps robustness: since the logic in the same one method is tested in all accesses, it gets much higher coverage.

Reviewed By: beauby

Differential Revision: D23566823

fbshipit-source-id: bd02e8217682fd5924194714b18de05c6492d9c1