Avoid overly-strict atomics in ringbuffer (#217)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/217

Disclaimer: I thought this would give us some (maybe small) perf gain, but it doesn't seem like that. Probably because x86 has a strong memory consistency model, and thus all atomics are very strict all the time. We might still see perf gains on other platform (think ARM), but we're not targeting them yet, so this is not really relevant. Regardless of the absence of perf gains, I still think this diff makes the code cleaner, as it specifies exactly the requirements we need.

We used atomics for the head/tail of the ringbuffers, which in their default behavior use a sequentially consistent memory model, which is the strongest possible one: all sequentially consistent accesses on any memory in the system are ordered in a single global total order. I expected this would cause some overhead (in the code, or in the processor). We don't need such a strong property though: we only need some ordering guarantees around the order of accesses within the ringbuffer, where we control all users. So we can express all the properties we need using the "weaker" acquire/release memory order.

See here for an explanation of memory orders: https://en.cppreference.com/w/cpp/atomic/memory_order

Reviewed By: beauby

Differential Revision: D23573343

fbshipit-source-id: 808e80d98d34eaa22fab14979bdcdc0a06419f6b