Rework SHM's ringbuffer logic
Summary:
The recent simplification of the ringbuffer interface allows, in turn, some simplification in its users, i.e., the connection. The read and write operations were in fact following very similar logic, but doing it in very different ways (e.g., the read operation had an `if(error)` after each call to the ringbuffer, and it would clean up and return early in each of them; the write operation instead had an `if(!error)` after each call, and would have a single cleanup at the end).

This diff removes some now obsolete code (handling impossible errors, spinlocking, ...), unifies the logic, making the read and write paths highly symmetric. It also tries to catch errors as early as possible (e.g., if the only "expected" error is ENODATA or ENOSPC, it filters that but throws on anything else), so that the logs we get are more informative.

Reviewed By: beauby

Differential Revision: D23575375

fbshipit-source-id: d8e6ef50b8410e8f812a04ffc4fb74960ac45bf8