Use RAII wrapper for mmapped memory
Summary:
We already had the shm::Segment behave as a RAII wrapper for its mmapped pointer, but it was also doing other things on top. Extracting the mmapped ptr to its own class I think helps readability and reusability. (And we plan to reuse the mmapped pointer in the InfiniBand transport). Note how now the Segment can get rid of its copy&move assignment/constructor and its destructor, because the default ones will do the job.

This diff also introduces a "pattern" for RAII objects, consisting of a class that wraps a unique_ptr with a custom deleter, implementing the acquisition of the resource in the constructor of the class and the release in the deleter. This then gives us a correct RAII class by delegating most of the logic to the unique_ptr, and highlighting the important logic. If we like such a pattern, I plan to use it more later on (e.g., for the ibverbs objects).

Reviewed By: beauby

Differential Revision: D23679003

fbshipit-source-id: 17276ceedb0ee327a0cdbe9d0d6f48ce475d730f