Handle scopes in IPv6 addresses in UV transport (#225)
Summary:
Contrary to IPv4, IPv6 could give the same link-local address to multiple interfaces. Moreover, while IPv4 uses link-local addresses only when it cannot find a global address, IPv6 always gives a link-local address to each interface. Thus link-local addresses need to be qualified with a sin6_scope_id field in the sockaddr_in6 struct, which in a textual representation becomes a `%<iface_name>` at the end of the IP address.

One issue we have is that the default way of auto-detecting the address in libuv (resolving the hostname and returning the first viable address) may find a link-local address. But then this is converted to string and back, and there we don't support the interface qualifier and thus lose it. When binding to a link-local address without an interface qualifier we get a EINVAL.

We could debate whether we should even support link-local addresses (and I think we will), but Gloo supports them (perhaps by accident, since Gloo doesn't convert to string but uses the sockaddr_in6 struct directly). We are aiming to mimic Gloo's behavior so, for now, let's support the interface qualifier in our string conversions.

The string-to-sockaddr one is easy because libuv has a function that can do it (and that function belongs to a family of functions, so for consistency I'll use them everywhere). The sockaddr-to-string conversion isn't automatically dealt with by libuv (on purpose it seems: https://github.com/libuv/libuv/issues/582) so I'm doing it manually.

I'm also adding tests for this which however can only run on Linux. We had other tests covering similar logic that were disabled, but I'm re-enabling them (also on Linux only) so that we can get at least some signal.

Fixes https://github.com/pytorch/tensorpipe/issues/224.

Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/225

Reviewed By: beauby

Differential Revision: D23958779

Pulled By: lw

fbshipit-source-id: 8f2970f47becbc8b365c5948d37372eeb49b0bad