Allow transports and channels to mark themselves as non-viable
Summary:
Some transports or channels may not be able to operate at all in a certain process or on a certain machine. For example, SHM may not be able to operate because /dev/shm isn't mounted (which is something that Docker does on purpose in some settings). CMA may not be able to operate because the Yama security module is preventing ptracing, or because the process_vm_readv syscall may be blocked (again something that Docker likes to do). CUDA_IPC may choose to be disabled because there are no CUDA devices. The upcoming IBV transport will want to remain disabled when no IB NICs are found, or when libibverbs isn't found.

To support all of this, this diff adds a isViable method on transport and channels, and then simply skips registering them if this evaluates to false.

Reviewed By: beauby

Differential Revision: D23078308

fbshipit-source-id: eda646a0568f1e9a2e6f6067720c4a773075489f