Add IP address auto-detection to IBV transport
Summary:
Even though InfiniBand isn't IP-based, it still uses a TCP socket to bootstrap the connection, exchange the info needed to set up IB queue pairs, and detect the remote end disconnecting. Therefore the user still needs to provide an IP address, both to listen on and to connect to. For our "regular" TCP transport (i.e., UV) we help the user figure out what a sensible address is by offering functions to resolve their own hostname or to get the address bound to an interface. For that, we follow the same logic as Gloo, for consistency. We should thus in turn provide the same for our IB transport, which is what this diff does.

Our users could, in practice, use the UV helpers to get a sensible address and then use that same address both for UV and for IBV. However, we'd like for our transports to be self-contained, as users could in principle build TensorPipe without UV but with IBV. And we cannot use UV's code in IBV, because we don't want IBV to depend on libuv. And we cannot use IBV's code in UV because it wouldn't be multi-platform. So, yes, we really need to duplicate this.

Reviewed By: beauby

Differential Revision: D24136727

fbshipit-source-id: 5622fd74427c9ad0d6b4f17514047607c14521b8