Mark IBV transport as non-viable when no IB devices found
Summary:
Auto-detection of the ability to connect two nodes is tricky in case of InfiniBand. Unlike SHM, where we can check if we're on the same machine, we cannot check whether we're on the same IB subnet. Unlike TCP, we cannot just assume that any two nodes can always connect, as IB subnets are typically small and isolated.

However, we're still going with that latter option because we're assuming that if the user is trying to connect two nodes that are part of the same training job and they both have access to an IB card, then they must be know what they are doing and probably want us to use IB.

This diff implements the "and they both have access to an IB card" part. Until now, the IB transport just crashed if it was used but no IB card was present.

Note that this check is somewhat brittle, as in principle all the methods of the transport are still accessible even if the check fails, and using them will likely cause a segfault. However, the way this check is used by the pipe is safe: if the check fails, the transport is immediately discarded and destroyed.

Reviewed By: beauby

Differential Revision: D24253193

fbshipit-source-id: 792addba18a8b74fa7587ac923b6414dc313821e