Add InfiniBand transport
Summary:
This diff introduces a transport for TensorPipe that leverages the InfiniBand verbs API. InfiniBand network cards (NICs) provide low-latency communication by allowing the user-space program to directly communicate with the card through memory-mapped I/O, thus side-stepping the kernel and avoiding the overhead introduced by syscalls and context switches. However, the NIC can only access memory that has previously been "registered" (i.e., pinned/page-locked), and its "physical" address has been given to the device. The buffers that users provide to TensorPipe aren't registered, and registering them is expensive (as it involves the kernel). There are also limits on the total amount of memory that can be registered.

The design we settled on thus uses some small pre-registered staging buffers, and then effectively performs three memory copies: from the source user buffer to the sender's staging area, from there (through InfiniBand) to the receiver's staging area, and finally from there to the target user buffer. Although that's a considerable number of memory copies, they are all performed in user-space or by the hardware, thus giving us an overall very good latency for small payloads.

The design, in fact, ends up resembling a lot the one of the SHM (shared memory) transport. There too we had a staging area (a ringbuffer), hosted in shared memory, that the two endpoints would copy data to/from. The main difference is that, while in SHM we have a single ringbuffer shared between the processes, in IBV we have two, one private to each endpoint, which are kept in sync through InfiniBand calls. (in fact, to support bidirectional transfers, an SHM connection has two buffers and an IBV one has four).

Also similarly to SHM, we can only achieve the desired low latency through a busy-polling loop (called reactor for historical reasons), as putting a thread to sleep and waking it up requires syscalls. Another similarity is that we need an out-of-band connection manager in order to bootstrap the connection: just like in SHM we first open a UNIX domain socket on which the two endpoints exchange the file descriptors of their shared memory regions, in IBV we first open a TCP socket to exchange the information needed to open a queue pair (InfiniBand's communication primitive). This TCP socket remains open for the whole lifetime of the connection but is never used for signaling, control or data transfers (as that would incur in a syscall). It is used to detect the other endpoint going away.

To further highlight the similarity with SHM, and to make reviewing easier, the files added in this diff have been marked as "copied" from the equivalent SHM ones (or, in some cases, the UV ones), which means that smart enough viewers will only display the delta between the IBV and the SHM ones, rather than marking all lines as new.

This diff still has several limitations that I'll address in subsequent diffs:
- Right now the IBV transport just crashes if it cannot find any device. We should instead have it mark itself as non-viable and thus be ignored by the pipe. However, we first need to add a way to do so.
- We need to cover the IBV transport in our CI. This is tough, as CircleCI doesn't provide machines with IB NICs. A solution would be to set up a virtual RXE device which basically simulates InfiniBand over IP in software. It would give us poor performance but probably be good enough to verify the correctness of our implementation.
- Ideally we'd want users who download a pre-built version of PyTorch to obtain a copy of TensorPipe built with the IBV transport. However, that would require users to have libibverbs.so installed, and PyTorch would crash if it's not the case. That's not acceptable. A solution could be to instead dlopen(...) libibverbs.so at runtime and, if that fails, again just mark the IBV transport as non-viable.
- The UV transport has some logic to help users figure out a reasonable default IP address to connect to (i.e., it resolves the hostname through DNS). We should add the same to IBV.
- It is highly recommended that ibverbs users also listen for async events on the ibverbs context, as some error conditions are only reported that way. We should do that.
- There's a possible optimization which, by allowing the sender's and receiver's ringbuffer to go out of sync (and thus adding complexity) could allow the sender to advance its ringbuffer's tail immediately when the RDMA write operation has flushed, instead of waiting for an ACK for that operation from the receiver. It's unclear what practical benefits this would bring.

Reviewed By: beauby, tscmoo

Differential Revision: D24136989

fbshipit-source-id: f1416d82266e36d8e48d9be8138eb76546727f77