Add helpers for InfiniBand verbs API: error checking macros
Summary:
Error checking macros for the three possible return values of ibverbs functions: an int (with -1 denoting error), a pointer (with nullptr denoting error) and void (which never fails, but we still use a macro so that the callsites don't look like we forgot to check for errors).

We add this code (and the one in subsequent diffs) as a file in the common/ subdir because it will be used both by the IBV transport and by the GDR channel.

Reviewed By: beauby

Differential Revision: D24136728

fbshipit-source-id: 7a295df9ad424978d331551e4b42b77246f9bb51