Add helpers to deal with dlopen/dlsym/...
Summary:
This diff is part of a stack that avoids linking with libibverbs at compile time, and instead opens the shared library at runtime. See D24253191 for more context.

In this diff, we add some RAII wrappers for dynamically loaded shared libraries, and a helper function to access symbols within it.

Reviewed By: beauby

Differential Revision: D24253190

fbshipit-source-id: bbdd561adb09f95ee43f9a7df3bf40650a1052b6