Support optional runtime linking with libibverbs
Summary:
This diff is part of a stack that avoids linking with libibverbs at compile time, and instead opens the shared library at runtime.

The reason to do this is that we want the official binary releases of PyTorch to be built with the IBV transport enable, but we want those binaries to also work on users systems that do not have libibverbs installed. In order words, we want libibverbs to become an optional dependency of the PyTorch binary.

Note that NCCL works in precisely the same way. See [here](https://github.com/NVIDIA/nccl/blob/master/src/include/ibvwrap.h) and [here](https://github.com/NVIDIA/nccl/blob/master/src/misc/ibvwrap.cc).

This diff introduces a class that opens and holds a handle to a dynamically-opened libibverbs.so, loads all the function symbols we care about and stores them as pointers, and provides access methods to conveniently invoke these functions.

Reviewed By: beauby

Differential Revision: D24253191

fbshipit-source-id: 3ab3fe02a49fb868bb6ba67d22f4db5ea06f1aeb