Re-implement ibverbs post and poll functions
Summary:
This diff is part of a stack that avoids linking with libibverbs at compile time, and instead opens the shared library at runtime. See D24253191 for more context.

In particular, this diff is the first of a series that tries to avoid even just including the libibverbs headers, by "replicating" them inside our own code. The reason for doing so is that, as we're hardcoding one version of the ABI (because we're explicitly loading `libibverbs.so.1`, note the trailing `.1`), we should use a matching API to then access and use the symbols we find in the shared library. In other words, during "standard" compilation and linking, the headers and the shared library go hand-in-hand and a consistent version of both of them is used for both phases: here we're trying to replicate that process but at runtime, so we should also make sure we use consistent API and ABI.

In practice I think this may not be likely to cause problems, because libibverbs is a stable and slow-moving library, and their [versioning guidelines](https://github.com/linux-rdma/rdma-core/blob/master/Documentation/versioning.md) are quite explicit about maintaining compatibility. But avoiding relying on their headers still feels safer to me.

Another advantage of not requiring their headers is that we also remove the dependency on libibverbs _at build time_. For PyTorch, for example, it may be tricky to ensure that the machines on which the releases are built do indeed have the libibverbs headers installed. If they don't, then the release would be built without IBV but it would be very hard to notice because the tests would still work as they would fall back onto UV. By not requiring the presence of the libibverbs headers in order to build IBV we can make sure that it's always built.

This diff defines some functions that are not present as symbols in libibverbs but that instead are defined inline in its headers.

Reviewed By: beauby

Differential Revision: D24331612

fbshipit-source-id: 0e8acd6998a7f3664a15dfcd14c291a8e0a4873e