Reproduce relevant sections of ibverbs headers in our own ones
Summary:
This diff is part of a stack that avoids linking with libibverbs at compile time, and instead opens the shared library at runtime. See D24253191 and D24331612 for more context.

This diff provides definitions for the relevant ibverbs data structures (structs, unions, enums, ...) inside our own headers, by replacing the aliases that we had introduced in the previous diff.

Reviewed By: beauby

Differential Revision: D24331610

fbshipit-source-id: 0d221f3166137bb31e117d58cb0789c0e78e0aea