Provide alias for all ibverbs data structures and constants in IbvLib
Summary:
This diff is part of a stack that avoids linking with libibverbs at compile time, and instead opens the shared library at runtime. See D24253191 and D24331612 for more context.

This diff prepares the ground to provide all the definitions of the ibverbs data structures (structs, unions, enums, ...) in our own headers. When we'll do that, these objects will have names nested within IbvLib. So in order to make that change smoother, we're introducing these names already, by making them aliases to the original ibverbs names, and fix all the rest of the code to use these new names. Thus no functional change happens. In the next diff, we'll replace these aliases with the real definitions.

Reviewed By: beauby

Differential Revision: D24331613

fbshipit-source-id: e39d7c04e36c8effec5313ed6144e9c5fa06a667