Make use of runtime-loaded libibverbs everywhere (#230)
Summary:
This diff is part of a stack that avoids linking with libibverbs at compile time, and instead opens the shared library at runtime. See D24253191 for more context.

In this diff, we use the helper class introduced in the previous diff throughout the IBV transport and the common helpers. All calls to the global `ibv_foo` functions are replaced with calls to the `foo` method of an `IbvLib` object. This object is stored on the reactor, as it's the object that already hosts all other long-lived ibverbs objects (context, protection domain, ...), and a reference to it is piped to every piece of code that needs to use it. The RAII wrappers, in particular, now need to also store a reference to this object in their deleter (in fact, though, they store a pointer, because the deleters must be default-constructible).

As, after this diff, all the libibverbs symbols that we need are retrieved at runtime, we also stop linking against libibverbs during compilation (although we still include its headers).

Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/230

Reviewed By: beauby

Differential Revision: D24253192

fbshipit-source-id: 1972449e0cfb9b21222988cd7186e1c4348d9e8f