Remove compile-time dependency on ibverbs from CMake (#231)
Summary:
This diff is part of a stack that avoids linking with libibverbs at compile time, and instead opens the shared library at runtime. See D24253191 and D24331612 for more context.

Now that we do not link with libibverbs nor do we make use of the definitions in its headers, we can in fact remove the dependency on it entirely. (We have in fact "simulated" a runtime dependency).

Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/231

Reviewed By: beauby

Differential Revision: D24331611

fbshipit-source-id: 4ffedf5b011f1b9f5b4f1c4bb2761d2b52375be0