Reorganize some includes
Summary:
Not a great title for this diff, but also not a great diff itself. There is one thing it tries to do, but also a bit of noise on top of it: until now, the definitions of the (abstract) transport and channel interfaces were structured by having listener.h, connection.h and channel.h forward-declare the Context, and the context.h import these files; after this diff the inverse happens, and context.h forward-declares Listener, Connection and Channel, while the other files include context.h.

The reason for that is that the complete definition of Listener, Connection and Channel is only needed within the project: end users don't have to have it. They do however have to include context.h, and the goal here is to set things to avoid them transitively including also the other ones.

One change induced by the above is that now any internal code that needs to _use_ a Listener, Connection or Channel needs to include the relevant header. One may think of this as overhead, but I also like the fact that it makes it very easy to find and track such users.

On top of this, this diff also tries to "optimize" a bit the includes inside the headers, especially those of the stdlib, by removing useless ones and adding those for classes that are referenced in the header (adding these includes isn't strictly necessary: the files were compiling, so apparently they were coming in transitively from somewhere else; but it's still nice to keep these files self-contained).

Reviewed By: patricklabatut

Differential Revision: D24416014

fbshipit-source-id: 03fe7cbc0d18cc07dcd330b0f247a510b359dd40